+++ tests/msm/imports/test_import.py
@@ -1,4 +1,7 @@
+from importlib.metadata import version
+
+
 def test_import_msm() -> None:
     import msm
 
-    assert msm.__version__ == "0.0.1"
+    assert msm.__version__ == version("ms-markets")